Better code reuse and renaming of files
